+++ ide/src/rust/ensogl/lib/components/src/list_view.rs
@@ -13,6 +13,7 @@ use ensogl_core::application;
 use ensogl_core::application::Application;
 use ensogl_core::application::shortcut;
 use ensogl_core::display;
+use ensogl_core::display::scene::layer::LayerId;
 use ensogl_core::display::shape::*;
 use ensogl_core::DEPRECATED_Animation;
 use ensogl_theme as theme;
@@ -411,6 +412,11 @@ impl ListView {
         frp.scroll_jump(MAX_SCROLL);
 
         self
+    }
+
+    /// Sets the scene layer where the labels will be placed.
+    pub fn set_label_layer(&self, layer:LayerId) {
+        self.model.entries.set_label_layer(layer);
     }
 }