+++ ide/src/rust/ensogl/lib/components/src/selector/decimal_aligned.rs
@@ -61,8 +61,8 @@ impl component::Model for Model {
         let label_full = app.new_view::<text::Area>();
         let label_left = app.new_view::<text::Area>();
 
-        label_full.remove_from_scene_layer_DEPRECATED(&app.display.scene().layers.main);
-        label_full.add_to_scene_layer_DEPRECATED(&app.display.scene().layers.label);
+        label_full.remove_from_scene_layer(&app.display.scene().layers.main);
+        label_full.add_to_scene_layer(&app.display.scene().layers.label);
 
         root.add_child(&label_full);
         root.add_child(&label_left);