+++ ide/src/rust/ide/view/graph-editor/src/component/breadcrumbs/project_name.rs
@@ -145,14 +145,14 @@ impl ProjectNameModel {
         text_field.set_default_text_size(text_size);
         text_field.single_line(true);
 
-        text_field.remove_from_scene_layer_DEPRECATED(&scene.layers.main);
-        text_field.add_to_scene_layer_DEPRECATED(&scene.layers.breadcrumbs_text);
+        text_field.remove_from_scene_layer(&scene.layers.main);
+        text_field.add_to_scene_layer(&scene.layers.panel_text);
         text_field.hover();
 
         let view_logger = Logger::sub(&logger,"view_logger");
         let view        = background::View::new(&view_logger);
 
-        scene.layers.breadcrumbs_background.add_exclusive(&view);
+        scene.layers.panel.add_exclusive(&view);
 
         let project_name = default();
         Self{app,logger,display_object,view,style,text_field,project_name}.init()