+++ ide/src/rust/ide/view/graph-editor/src/component/node/output/area.rs
@@ -179,8 +179,8 @@ impl Model {
         // FIXME[WD]: Depth sorting of labels to in front of the mouse pointer. Temporary solution.
         // It needs to be more flexible once we have proper depth management.
         let scene = self.app.display.scene();
-        self.label.remove_from_scene_layer_DEPRECATED(&scene.layers.main);
-        self.label.add_to_scene_layer_DEPRECATED(&scene.layers.label);
+        self.label.remove_from_scene_layer(&scene.layers.main);
+        self.label.add_to_scene_layer(&scene.layers.label);
 
         let text_color = self.styles.get_color(theme::graph_editor::node::text);
         self.label.single_line(true);